successfully created delete user listing functionality

+++ API/controller/listing.controller.js
@@ -1,24 +1,37 @@
-import listing from "../../Models/ListingModel.js"
-import { errorHandler } from "../utils/error.js"
+import listing from "../../Models/ListingModel.js";
+import { errorHandler } from "../utils/error.js";
 
-export const createListing = async (req,res,next) => {
-    try{
-     const list = await listing.create(req.body)
-     res.status(201).json(list)
-    }catch(error){
-        next(error)
+export const createListing = async (req, res, next) => {
+  try {
+    const list = await listing
+      .create(req.body)
+      .then((us) => {
+        console.log("Listing created successfullt: ", us);
+        res.status(201).json(us);
+      })
+      .catch((err) => {
+        console.log("error occured");
+      });
+  } catch (error) {
+    next(error);
+  }
+};
+export const deleteListing = async (req, res, next) => {
+  const listings = await listing.findById(req.params.id);
+  let indexToRemove = Number(req.params.index);
+  if (!listings) {
+    return next(errorHandler(404, "Listing not found"));
+  }
+  try {
+    if (listings.imagesUrls.length === 1) {
+      await listing.findByIdAndDelete(req.params.id);
+      res.status(200).json("Listing has been deleted");
+    }else{
+        listings.imagesUrls.splice(indexToRemove, 1);
+        await listings.save();
+        res.status(201).json(listings)
     }
-}
-export const deleteListing = async (req,res,next)=>{
-    const listings = await listing.findById(req.params.id)
-
-    if(!listings){
-        return next(errorHandler(404, 'Listing not found'))
-    }
-    try{
-       await listing.findByIdAndDelete(req.params.id)
-       res.status(200).json('Listing has been deleted');
-    }catch(error){
-        next(error)
-    }
-}
+  } catch (error) {
+    next(error);
+  }
+};
